Move code for DashboardLinkOrButton cmp to own file.
